+++ r_bsp.c
@@ -76,8 +76,8 @@ void R_RotateBmodel (void)
 	// yaw
 	angle = currententity->angles[YAW];
 	angle = angle * M_PI*2 / 360;
-	s = sin(angle);
-	c = cos(angle);
+	s = sinf(angle);
+	c = cosf(angle);
 
 	temp1[0][0] = c;
 	temp1[0][1] = s;
@@ -93,8 +93,8 @@ void R_RotateBmodel (void)
 	// pitch
 	angle = currententity->angles[PITCH];
 	angle = angle * M_PI*2 / 360;
-	s = sin(angle);
-	c = cos(angle);
+	s = sinf(angle);
+	c = cosf(angle);
 
 	temp2[0][0] = c;
 	temp2[0][1] = 0;
@@ -111,8 +111,8 @@ void R_RotateBmodel (void)
 	// roll
 	angle = currententity->angles[ROLL];
 	angle = angle * M_PI*2 / 360;
-	s = sin(angle);
-	c = cos(angle);
+	s = sinf(angle);
+	c = cosf(angle);
 
 	temp1[0][0] = 1;
 	temp1[0][1] = 0;
