+++ r_main.c
@@ -267,7 +267,7 @@ void R_ViewChanged (vrect_t *pvrect, int lineadj, float aspect)
 	if(ow != r_refdef.vrect.width || oh != r_refdef.vrect.height)
 		vid.recalc_refdef = true;
 
-	r_refdef.horizontalFieldOfView = 2.0 * tan (r_refdef.fov_x/360*M_PI);
+	r_refdef.horizontalFieldOfView = 2.0 * tanf(r_refdef.fov_x/360*M_PI);
 	r_refdef.fvrectx = (float)r_refdef.vrect.x;
 	r_refdef.fvrectx_adj = (float)r_refdef.vrect.x - 0.5;
 	r_refdef.vrect_x_adj_shift20 = (r_refdef.vrect.x<<20) + (1<<19) - 1;
@@ -353,7 +353,7 @@ void R_ViewChanged (vrect_t *pvrect, int lineadj, float aspect)
 	for (i=0 ; i<4 ; i++)
 		VectorNormalize (screenedge[i].normal);
 
-	res_scale = sqrt ((double)(r_refdef.vrect.width * r_refdef.vrect.height) /
+	res_scale = sqrtf((float)(r_refdef.vrect.width * r_refdef.vrect.height) /
 			          (320.0 * 152.0)) *
 			(2.0 / r_refdef.horizontalFieldOfView);
 	r_aliastransition = r_aliastransbase.value * res_scale;
@@ -756,8 +756,8 @@ void R_InitTurb (void)
 	int i;
 
 	for(i = 0; i < SIN_BUFFER_SIZE; i++){
-		sintable[i] = AMP + sin(i*M_PI*2/CYCLE)*AMP;
-		intsintable[i] = AMP2 + sin(i*M_PI*2/CYCLE)*AMP2;	// AMP2, not 20
+		sintable[i] = AMP + sinf(i*M_PI*2/CYCLE)*AMP;
+		intsintable[i] = AMP2 + sinf(i*M_PI*2/CYCLE)*AMP2;	// AMP2, not 20
 	}
 }
